Oops, that fix actually broke it.  Revert.
Here's what was going on:  In testing, my email appeared as ryan@dovecot.  When I would generate a key with that id, I got an error saying that it was an invalid user id.  I thought, "the authors of openpgpjs wouldn't have done something as ridiculous as apply a regex to the email address to check if it had a tld or something!  The error must be somewhere in rc_openpgpjs."  So I found what looked like an error and fixed it.

But it turns out the problem does reside somewhere in rc_openpgpjs.  They really do have a regex that attempts to detect "invalid" email addresses.  I cannot fathom why.  They do not envision that a pgp setup may be used in an all-internal environment, where it is valid to use a hostname other than one from the public DNS from ICANN.  Is this in some sort of PGP spec?  Or is it just the shortsightedness that seems to take hold of everybody when someone mentions email addresses and they say "Oh, I know what an email address looks like!  I better enforce my preconceived notions, even though nothing would go wrong if I didn't!"

The function in openpgpjs is "isEmailAddress", in util.js.  I should fix it.