Key selector shows all persons.
If an individual key had N userids associated with it, we used to just show one of the userids, but show it N times.  Now, we show all the userids.

I'd be happier if we had a table like the one in the key manager, so it was more obvious that all the userids were for the same key.  But this works.