Don't stop showing spinner until encryption is done.
In the key select dialog, when the user has entered a passphrase, we want to hide the submit button and show "working" until we're done encrypting, because it might take a second or two.  Previously, we removed the spinner after decrypting the private key, which was too soon.

However, even now, I never see the spinner.  I don't know why.  If I add a 3sec delay before starting the encryption, I do see the spinner.