Switching key select to use keyId
Previously, it was using the first eight chars of the fingerprint, which is not used anywhere else in openpgp or gpg, so it would not be useful.  The keyId is something that is used.  By default, command-line gpg shows the "short format" keyId, and that's a more human-digestible string.  I wish we could use that here, but openpgp's getKeyId, but openpgpjs does not do the short format.

Another important reason to switch to keyId is for the "Remember for this session" feature of the key selector.  Previously, what we'd been remembering was the position of the key in the array of all private keys.  That means that if the user selected a key and remembered it, and then went to the key manager and deleted an earlier key from the keyring, then that position index is no longer valid, and points to a different key.  When we remember, we also remember the passphrase.  The remembered passphrase probably doesn't work on the new key, and if it does, that's worse because the user will think we've been remembering one key, and we're actually using a different one!

To overcome this problem, we will remember the keyId and passphrase.

Now, we could simply remember the ascii-armored version of the decrypted private key.  After all, the keyId and passphrase can be used to obtain the decrypted private key.  But it just made me nervous to have that in memory between sessions.

So instead, we will remember keyId and passphrase.  We will use the keyId to look up the key later, when we want to use it for signing.  And that means we need an id we can look things up by.  The choices listed by openpgpjs's getKeysForId function are: keyId or full fingerprint.  First eight chars of the fingerprint cannot be used to look anything up.

So, the state that the code is in right now:
The key selector shows the keyIds.  Clicking on a key will show it.  That's all I've tested since changing this, and that's probably all that works.  Let's go change the rest of it, so it's all about keyId!