Only decrypt the private key once.
Previously, we decrypted the private key twice:  Once to see if the passphrase was correct.  Then we would throw that decrypted key away, and then decrypt it a second time, when we were about to sign the message.

It was easy to do this because, once we found that the passphrase worked, the object we passed back contained a position index.  We would use that index to look up the key in the keyring.  We used to do that because it was easy to store the position index in the sessionStorage, if the user asked to use the same key and remember the passphrase each time.  That meant we could have one code path for both situations: 1) a key selected just now, 2) a key pulled from sessionStorage.

But, for reasons already explained in an earlier commit log, the position index was not a reliable way to identify the key.  (What if the keyring changed between saving the preference and using it?)  So I'm changing it so that the sessionStorage stores the keyId instead of a position index.  The logic to get the decrypted key from the keyId requires sufficient library-specific code that I would want to put it in rc_openpgpjs_crypto instead of rc_openpgpjs.  But it seemed silly to do that only to require us to decrypt the key a second time, in the case of a key selected just now.

For a key pulled from sessionStorage, we'll just have a different code path that decrypts it. (not implemented yet)

(NOTE: I should test what happens if the key gets deleted from the keyring between saving the preference and using it!  We should detect that case and pop up the key select dialog, but we probably don't.)