Remember key selection/passphrase for the session.
You can click the checkbox to remember the key selection and passphrase for this session.  That will work.

Now, the keyring may have changed between the time when you remembered the preference and the time when you use it to send a message.  If the key is deleted or the passphrase changed, then you won't be able to decrypt the key.  In that case, we simply present the key selection dialog again.