We should have a UI to forget remembered keys.
This commit adds some HTML to the compose messages interface.

In the current UI, when you are about to send a signed message, you're given a choice of what key to sign with, and you're made to enter the passphrase.  You may want to remember this selection, and remember the passphrase so you don't have to enter it again next time.  We have a checkbox on the form when you enter your passphrase.  If you click the checkbox, the passphrase and selection are saved for the remainder of the browser session.

However, what if you change your mind and want to start using a different key?  What if you're about to hand your computer to someone else and don't want them to be able to sign messages without having to enter your passphrase?  Currently, there is no way to forget your selection or your passphrase without closing your browser (or something; I'm not entirely clear on how sessionStorage is supposed to work).

The idea is that, if you have a remembered key, it should show a UI element showing you that you have a key remembered.  That way, you won't click send thinking that it'll give you a choice of private key when actually, it remembered a previous selection.  If you do that, you may sign a message with a private key you did not intend to.  Maybe you didn't want people to know that those private keys belonged to the same person.  Well, this UI should let you see that this will happen.  If you make that mistake, then - hey - at least you were warned.

Also, when you see what key is remembered, you'll have the option to forget the key by clicking a button.  If you do that, then next time you try to sign a message, you'll get the key selection / passphrase entering dialog again.

Okay, but I just noticed something:  Maybe you want to remember the passphrase for a bunch of different private keys, because you are reading encrypted messages.  Remembering those passphrases should be different from remembering what key you want to be the default when you sign a message.  How should we separate these ideas in the UI?  I don't know yet, but think about it.